backend/onyx/caseprediction/requirements.txt
fastapi
uvicorn>=0.22.0
transformers==4.39.3
torch
langchain-community>=0.2.1
python-dotenv
pydantic
chromadb
authlib
beautifulsoup4
markdown
requests
PyYAML
huggingface-hub>=0.23.4
sentence-transformers==2.6.0
pypdf
aiohttp
faster-whisper
google-generativeai
PyJWT
pyjwt[crypto]
passlib[bcrypt]
uuid
apscheduler
tokenizer==3.4.5
sentencepiece

